LLM_Project/requirements.txt
torch>=2.2
torchvision>=0.17
pillow>=10.0
numpy>=1.24
tqdm>=4.66
tensorboard>=2.14
fastapi>=0.110
uvicorn>=0.29
scikit-learn>=1.3
jupyter>=1.0
